Remove post processing step from parser pipeline of `filestream` (#25243)
## What does this PR do?

This PR removes the post-processing done by JSON parser. It has become possible because I moved adding `log` metadata to a `reader.Reader`. Thus, JSON parser can overwrite these values if necessary.

The PR also adds support for multiple parsers of the same type. It is not the best solution, so I am open to alternatives. At the moment JSON parser generates an ID for itself. The parsed keys are added to `Fields` under the key `"json{key}"`. The problem with this solution that the event looks somewhat ugly:

```json
{
    "@timestamp": "{my-timestamp}",
    "json1234-1234-1244": {
        "key": "value"
    }
}
```

## Why is it important?

Fixes the abstraction leak in `filestream` input.