Ignore https errors (#27412)
Fixes #27202

NOTE: This requires a version of synthetics with elastic/synthetics#361

Adds an option to disable errors on invalid TLS certificates in heartbeat.

Rather than try to work with ssl.verification_mode: none this adds a new option. The reason being that the behavior here is just too different to share a configuration option.