[Elastic Agent] Disable modules.d in metricbeat (#27860)
Currently when Elastic Agent is started, on post-install of metricbeat the system.yml file in the modules.d directoy is renamed to system.yml.disabled to make sure it is not run. It turns out there might be some edge cases where this does not work. To not rely on renaming of files, instead the modules are fully disabled.

Closes https://github.com/elastic/beats/issues/27857