Initialise beat early
Initialise beat early so all required information for monitoring and
state endpoint are available when the state registy is populated.

A unit test is added and the mockbeatk (libbeat/mock/mockbeat.go) is
updated to make it explicity that `mock.New` does not use any of its
input parameters.

Fixes #27755