Add metadata change support for processors (#30183)
Some of the processors that are configured with a target field didn't
support special cases with the target field is a `@timestamp` or a
`@metadata` sub-field. This change is to make it consistent across
processors or to document why the processor does not need this change.

Also added tests for each processor that is supposed to support metadata as a target.