Add a ring-buffer reporter to libbeat (#28750)
* Add a ring-buffer reporter to libbeat

Add a ring buffer reporter that when enabled will store configured
namespaces in a buffer to allow operators to view recent metrics
history. Defaults are to gather the stats namespace every 10s for 10m.
Must be explicitly enabled, along with monitoring, and the HTTP
endpoint. The buffer endpoint is intended to be used for diagnostics
reporting.

* Add basic benchmarks

* Review feddback