Remove prev-minor testing as not needed anymore (#30926)
In https://github.com/elastic/beats/issues/26629 the issue around running Filebeat against older version of Elasticsearch was discussed and in https://github.com/elastic/beats/pull/28274 testing against the previous minor was introduced. But since 8.0, Filebeat can only ship data to equal or newer versions of Elasticsearch. Because of this, in the tests `TESTING_FILEBEAT_ALLOW_OLDER=1` had to be introduced.

I'm opening this PR to remove this testing as it is not something we support anymore.