[Discuss] Update Codeowners file (#30793)
The following is a detailed update of the codeowners file for the different teams. Each top level directory is listed with the team that owns it. In some cases like modules more specific details were added. Every time we see some specific directories belong to a team, the Codeowners file should be updated.

The goal of this PR is to have alignment across all the teams on ownership.

Few additional notes:

* The robots team is not an owner of any files / directories. This is by design as the final ownership of the build is with the data plane team.
* Modules are owned by default by the integrations team. More details must be added here.
* system tests are not covered in details

Co-authored-by: Craig MacKenzie <craig.mackenzie@elastic.co>