packetbeat/config: allow multiple interfaces to be specified in the config (#32732)
Currently only the first interface is used, retaining the existing behaviour. If
command line arguments are provided the existing behaviour is kept with the first
interface having options from the command line replacing being placed in that
config as defaults, replaced by the options in the config file. Duplicated
interface device declarations are disallow in the configuration.