auditbeat/module/file_integrity: add file parsing processor (#28802)
This adds a file parser and analyser framework along the lines of the approach
described in elastic/beats#18048. The processor allows inspection of files visited
during a scan or seen in events based on a document path configuration.

A file analyser is also included that performs an analysis of executable or library
object files for a variety of platforms, providing details of imports and other
potential threat indicators.