winlogbeat/sys/winevent: do not interpret dots as JSON path separators (#34549)
In some locales, Windows abbreviates event data keys with dots. This
happens in Spanish for example where the English 'ID' is rendered as 'Id.'
When using mapstr.Put, this dot is interpreted as a path separator,
breaking the structure of the document and ultimately leading to an
ingest failure. So use opaque key map insertion.