Fix system module with both filesets enables
The system module did not define an ID at the root of the config, that
made the V2 input loader only start the first journald input it saw
because they both ended up with the same identifier (type, ID and
path). This is fixed by defining an ID at the root of the
configuration templates.

The journald input now also adds the `input_id` key to its loggers
and a non-fatal error is now logged at debug level.

The system-logs input is now marked as experimental instead of stable.